braces to avoid ambiguous ‘if if else'

+++ common/pictureflow.cpp
@@ -1259,18 +1259,18 @@ void PictureFlow::showSlide(unsigned int index)
 {
   index = qMax<unsigned int>(index, 0);
   index = qMin<unsigned int>(slideCount()-1, index);
-  if(index == d->state->centerSlide.slideIndex)
-    return;
+  if(index == (unsigned int)d->state->centerSlide.slideIndex)
+      return;
 
-    int distance = centerIndex()-index;
+  int distance = centerIndex()-index;
 
-    if(abs(distance)>10)
-    {
-        if(distance<0)
-            setCenterIndex(centerIndex()+(-distance)-10);
-        else
-            setCenterIndex(centerIndex()-distance+10);
-    }
+  if(abs(distance)>10)
+  {
+      if(distance<0)
+          setCenterIndex(centerIndex()+(-distance)-10);
+      else
+          setCenterIndex(centerIndex()-distance+10);
+  }
 
   d->state->centerIndex = index;
   d->animator->start(index);
